# ebpf: optimize SysEnter and SysExit probes