@NDStrahilevitz 2022-08-29
> It looks good to time. As another optimization I would skip all timing related syscalls:
> 
> ![image](https://user-images.githubusercontent.com/7395852/187228693-48640028-7f8e-4399-a3e9-2b8804bea287.png)
> 
> right away (without even doing map lookups). @yanivagman mind taking a look at this change ? I think it is solid, but the code is sensitive and getting a second look would be good.

Sounds like a good additional optimization, maybe for another PR rather than in this one though. I'll add it to the the issue tracking optimization.