@NDStrahilevitz 2022-08-29
> > It looks good to time. As another optimization I would skip all timing related syscalls

Sounds like a good additional optimization, maybe for another PR rather than in this one though. I'll add it to the the issue tracking optimization.