@yanivagman 2022-08-29
> It looks good to time. As another optimization I would skip all timing related syscalls

What if the user chose to trace one of these syscalls?