@rafaeldtinoco 2022-08-30
> > It looks good to time. As another optimization I would skip all timing related syscalls
> 
> What if the user chose to trace one of these syscalls?

Well I did not think we had that use case but I see your point. And thinking a bit more about the suggestion, I believe a big optimization for cases relying in timing syscalls a lot is to use vDSO (what glibc and musl might be already doing), so... nevermind. =)